Update CMake setup for MetaCG
This is motivated by the need to make the MetaCG graph lib usable for third party tools and has several larger changes:
1. CGCollector and PGIS are *no longer built by default*
2. An installation of MetaCG generates files necessary for CMake's `find_package` and `add_link_libraries` to work.
3. nlohmann::json and spdlog are installed when installing MetaCG.
4. CMake variables to enable the build of CGCollector and PGIS have been renamed to `METACG_BUILD_<NAME>` to make it clear what happens.

As a result of 2: MetaCG's graph lib can be linked using `add_link_libraries` with the namespaced library `metacg::metacg`. This pulls in all flags / libraries etc.
To make this happen, the nlohmann::json library and the spdlog library are installed as part of the MetaCG installation.
This happens only if they are not pointed to by using `METACG_USE_EXTERNAL_JSON` and `METACG_USE_EXTERNAL_SPDLOG`.