[PGIS] Starting to clean up CLI options
Cleans up the CLI options somewhat as mentioned in #90.

The major difference is that every option option is stored in the global config (as done before, but now less convoluted) and the option's value is now returned instead of set through a passed-in reference. This enables us to get rid of a good number of overly useless local variables in `main` and makes the code cleaner.