Adds the new, overhead-based heuristic described in JR's thesis "Overhead-Guided Instrumentation Refinement"
The core idea of the overhead-based heuristic is to adjust the number of instrumented function calls, based on the observed overhead (provided via the argument "prev-overhead") and the target overhead (provided via the argument "target-overhead").
If the observed overhead is lower than the targeted overhead limit, the number of instrumented calls is increased, otherwise lowered.
It is assumed that each instrumented function generates approximately the same overhead
The number of calls to a function are estimated based on code structure info extracted by cgc or, if available, previous instrumentation results are used.
To select functions to instrument a greedy algorithm is used to solve the knapsack problem, with the weight/cost being the (estimated) number of calls to a function, representing its estimated instrumentation overhead, and the value being the inclusive statement count, representing the estimated information gain.
Removing functions from the instrumentation is done based on the observed runtime in relation to the number of invocations. Different function removal algorithms can be selected by the parameter "overhead-selection".
The cost of instrumenting (likely) inlined and recursive functions can be scaled, as they often have an over-proportional high instrumentation overhead.

The new overhead-based heuristic uses new metadata to store instrumentation results and decisions over multiple PIRA iterations

A new estimator phase can be used to, optionally, fill gaps in the instrumentation