[CMake][CI] Fix formatting and CI errors
This factors out the policy settings to a module and function so it can
be included across the project. Also, this allows to run cmake-format
again on the CMakeLists.txt file without crashes.

Don't know why there were so many failures in the CMake formatting, but
anyway, this patch fixes the CMake formatting.